refactor: wip, pr comments
